dev-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile dev-requirements.in
#
attrs==22.1.0
    # via pytest
exceptiongroup==1.0.0
    # via pytest
faker==12.0.1
    # via mixer
flake8==5.0.4
    # via
    #   -r dev-requirements.in
    #   flake8-isort
    #   flake8-pyproject
    #   flake8-walrus
flake8-isort==5.0.0
    # via -r dev-requirements.in
flake8-pyproject==1.1.0.post0
    # via -r dev-requirements.in
flake8-walrus==1.2.0
    # via -r dev-requirements.in
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via flake8-isort
mccabe==0.7.0
    # via flake8
mixer==7.2.2
    # via -r dev-requirements.in
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r dev-requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r dev-requirements.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via
    #   flake8-pyproject
    #   pytest
